requirements.txt
pandas>=2.0
numpy>=1.26
python-dotenv>=1.0
tqdm>=4.66
openai>=1.40
pymupdf>=1.24
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
gradio>=6.0
langchain-core>=0.3
langchain-community>=0.3
langchain-openai>=0.2
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3
beautifulsoup4>=4.12
